@@ -1,10 +1,30 @@
-import Vue from 'vue'
+// import Vue from 'vue'
 import App from './App.vue'
-import router from './router'
+import {createRouter, createWebHistory} from 'vue-router'
+import MinuteHero from "@/components/minuteHero/MinuteHero";
+import {createApp} from 'vue'
 
-Vue.config.productionTip = false
+const routes = [
+  {path: '/minute-hero', component: MinuteHero},
+  {path: '/', component: MinuteHero},
+  {path: '', component: MinuteHero},
 
-new Vue({
-  router,
-  render: h => h(App)
-}).$mount('#app')
+]
+//
+const router = createRouter({
+  history: createWebHistory(),
+  routes, // short for `routes: routes`
+})
+
+const a = createApp(App);
+a.use(router);
+// a.use(Toast, options);
+// a.component('fa-icon', FontAwesomeIcon)
+a.mount("#app");
+
+
+// Vue.config.productionTip = false
+// new Vue({
+//   router,
+//   render: h => h(App)
+// }).$mount('#app')
