@@ -1,38 +1,34 @@
 <template>
-	<div id="app">
-		<div id="nav">
-			<!--			<router-link to="/">Home</router-link> |-->
-			<!--      <router-link to="/vroomsticks">Vroomsticks</router-link> |-->
-			<!--			<router-link to="/smashketball">Smashketball</router-link> |-->
-			<router-link to="/">The Witch is Dead!</router-link>
-			<!--      <router-link to="/about">About</router-link>-->
-		</div>
-		<router-view/>
-	</div>
+  <div id="app">
+    <div id="nav">
+      <router-link to="">2 Minute Hero</router-link>
+    </div>
+    <router-view/>
+  </div>
 </template>
 
 <style>
 #app {
-	font-family: Avenir, Helvetica, Arial, sans-serif;
-	-webkit-font-smoothing: antialiased;
-	-moz-osx-font-smoothing: grayscale;
-	text-align: center;
-	background-color: #2c3e50;
-	color: aliceblue;
-	height: 100vh;
-	width: 100vw;
+  font-family: Avenir, Helvetica, Arial, sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  text-align: center;
+  background-color: #2c3e50;
+  color: aliceblue;
+  height: 100vh;
+  width: 100vw;
 }
 
 #nav {
-	padding: 30px;
+  padding: 30px;
 }
 
 #nav a {
-	font-weight: bold;
-	color: #42b983;
+  font-weight: bold;
+  color: #42b983;
 }
 
 #nav a.router-link-exact-active {
-	color: #42b983;
+  color: #42b983;
 }
 </style>
